requirements.txt
# Minimum requirements
#
rnbgrader>=0.1.1
python-statemachine
